requirements.txt
flask
flask_restful
flask_sqlalchemy
flask_jwt_extended
requests
passlib
pillow